refactor: enhance wallet switcher and add arns gating
